Added PolygonWith2Heads3D shape.
Added PolygonWith2Heads3D shape. This shape consists of a 2d polygon with 2 vertices connected to its vertices, one "in front" and one "from the back". With this shape we can create a double pyramid shape (already did this one), a diamond or any similar shape with wanted vertices, radius and height.