Added drawing POINTS using mouse option.
Added the option to draw points using left mouse clicks. the number of points that can be drawn in the same instance is limited using the public variable drawingCapacity, when we exceed this number we start replacing the earliest drawn points with new ones.
I used the linked list data structure to maintain the points.

next step: allow drawing lines.