Added PointManager and LineManager data structures.
PointManager:
deals with adding points and removing them, also deals with checking if a given point intersects with an already existing point. This class is also responsible for drawing the points.

Line Manager:
deals with adding lines and removing them, it also deals with deleting all the lines that connect with a point that we delete. This class is also responsible for drawing the lines.

THE SOLUTION IS NOT IMPLEMENTED YET - FOR NOW THE CODE ONLY DRAWS POINTS AND LINES TO THE MIDDLE POINT - THE "HOLD LEFT MOUSE BUTTON" LOGIC IS NOT IMPLEMENTED YET

THE CODE STRUCTURE IS BAD - will fix later