Fixed "line not fully drawing while holding down" during runtime.
The issue: when we hold down the mouse button and move the mouse, the line we draw looks like it "doesn't have enough time to draw" and so it keeps cutting off like a stream of water being blocked and opened really fast.
Turns out that the actual issue is that when we call the DrawLine function inside OnLeftMouseHold, we hand over the following HDC: GetDC(hwnd), which in fact hands us the actual HDC and not the memDC for the bitMap.
I fixed this issue by adding the HDC parameter to the function and passing in memDC in WinMain.