Added moving lines logic.
We can now move lines by pressing MMB or 'M', holding, and letting go.

only lines... for now.
Next step:
Create a common manager between LineManager and PointManager.

For this feature, we can feed the new LineManager function getLinesWithPoint() with the Point we get from running CheckInterseciton() in PointManager, this will yield the point and the lines we want to move.