Fix Line length not being drawn bug.
Fixed the bug where the length of the line doesn't get drawn. I don't know how long this bug has been lurking. The problem was with the copy constructor of Line, I was copying the lines somehow but the copy constructor of Line doesn't copy the length and the Point "Middle".