Add class PointLineManager. Fix CheckIntersection function logic. Find minor issue (still Unfixed)
Added the class PointLineManager just how I explained in TODO List.txt.

As I was writing the code, I ran into a minor issue: when we call PointManager::CheckIntersection, if we find a point in range we return it, and if we didn't find any point we return the original point. This logic is flawed, considering that we may CheckIntersection with a point of exact coordinate as that in the data structure, and returning the point in the data structure will leave us with no possible way of knowing whether we didn't find any point in range or that the exact point already existed.

Thus, I solved this using std::optional, if we found a point in range, we return it, else we return std::nullopt.

This was my first time dealing with std::optional, it feels illegal.

IMPORTANT NOTE:
After adding the moving points and lines feature, I was greeted with another small bug: because I took the lines as static when I added the length text feature, when we move a line, the text draws at the old spot.
I introduce the solution in the TODO List.txt file.