Change Animation logic. Add Lerp class.
Changed the logic around animations, instead of defining an animation and giving it a function to run on its thread, I changed the logic to have the Animation class as an abstract class, and now when we want to define an animation, we can define a class for that animation and make it inherit from the animation class, and now it can implement its own DoAnimation function.

Added lerp class: takes in two points to linearly interpolate between and a duration dur, it makes an animation of a point going from point a to point b in dur seconds.