Finish Animator class. Fix Lerp class. Integrate Time and Animation logic in main.
Finished Animator class: we can now freely add animations and it should all work as intended.

Fixed Lerp class. now we draw original point, and the last point we wanna get to, and a line between the original and the current points, can change if we want, but I like it this way.

Integrated Time and Animation logic in main. Added the clock and animator objects, we get the time related stuff from clock and add and call animations from animator.